+++ src/app/app.component.ts
@@ -1,13 +1,12 @@
 import { AfterViewInit, Component, ElementRef, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Subject } from 'rxjs';
-import { takeUntil, tap } from 'rxjs/operators';
-import { CartService } from './cart/services';
-import * as CartActions from './cart/store/cart.actions';
 import { LocalStorageService } from './core/services';
 import { AppState } from './reducers';
 import { Roles } from './shared/models/roles';
 import { UserDataModel } from './shared/models/user-data.model';
+import * as CartActions from './cart/store/cart.actions';
+import * as ProductsActions from './products/store/products.actions';
 
 @Component({
   selector: 'app-root',
@@ -29,12 +28,11 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
   constructor(
       private storageService: LocalStorageService,
       private store: Store<AppState>,
-      private cartService: CartService,
   ) {}
 
   ngOnInit(): void {
       this.storageService.setItem('currentUser', this.adminUser);
-      this.initCart();
+      this.initStore();
   }
 
   ngAfterViewInit(): void {
@@ -46,12 +44,8 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
       this.takeUntil$.complete();
   }
 
-  private initCart(): void{
-      this.cartService.getCart().pipe(
-          takeUntil(this.takeUntil$),
-          tap((items) => {
-              this.store.dispatch(CartActions.addItems({items}));
-          })
-      ).subscribe();
+  private initStore(): void{
+      this.store.dispatch(CartActions.getCart());
+      this.store.dispatch(ProductsActions.getProducts());
   }
 }