+++ src/app/cart/store/cart.actions.ts
@@ -5,17 +5,22 @@ export const getCart = createAction(
   '[Cart] Get Cart'
 );
 
+export const getCartSuccess = createAction(
+    '[Cart] Get Cart Success',
+    props<{items: CartItemModel[]}>(),
+);
+
 export const emptyCart = createAction(
     '[Cart] Empty Cart'
 );
 
-export const addItems = createAction(
-  '[Cart] Add Items',
-  props<{items: CartItemModel[]}>(),
-);
-
 export const addItem = createAction(
   '[Cart] Add Item',
+  props<{item: CartItemModel}>(),
+);
+
+export const addItemSuccess = createAction(
+  '[Cart] Add Item Success',
   props<{item: CartItemModel}>(),
 );
 
@@ -24,14 +29,29 @@ export const removeItem = createAction(
   props<{id: number}>(),
 );
 
+export const removeItemSuccess = createAction(
+    '[Cart] Remove Item Success',
+    props<{id: number}>(),
+);
+
 export const increaseQuantity = createAction(
     '[Cart] Increase Quantity',
-    props<{id: number}>(),
+    props<{item: CartItemModel}>(),
+);
+
+export const increaseQuantitySuccess = createAction(
+    '[Cart] Increase Quantity Success',
+    props<{item: CartItemModel}>(),
 );
 
 export const decreaseQuantity = createAction(
     '[Cart] Decrease Quantity',
-    props<{id: number}>(),
+    props<{item: CartItemModel}>(),
+);
+
+export const decreaseQuantitySuccess = createAction(
+    '[Cart] Decrease Quantity Success',
+    props<{item: CartItemModel}>(),
 );