+++ src/app/products/store/products.actions.ts
@@ -7,6 +7,10 @@ export const getProducts = createAction(
 
 export const setProducts = createAction(
   '[Products] Set Products',
+);
+
+export const getProductsSuccess = createAction(
+  '[Products] Get Products Success',
   props<{products: ProductModel[]}>(),
 );